Board config and initial test
An Adafruit PN532 breakout board is connected to a Photon using I2C.  The firmware runs a basic test that connects to the breakout board, reads out its firmware version and awaits for an RFID card to be presented.  Whn RFID card is present, the UID is read out, tested to see if it is a Mifare Classic card and if so, the UID is decoded and a message is printed.