Update test-PN532-keyChange.ino
This code adds functions to authenticate a block (with relative block numbers and error checking) and to read data, including authenticating a block.

This version also has a function to determine the card type based upon trying to authenticate with various keys.  HOWEVER, it appears that if you authenticate with a bad key and it fails, then authenticating with a good key will also fail!