Update test-PN532-keyChange.ino
Fixed the testCard() function so that it resets after authentication failure and also only tries key A for authentication (so that it can be used in any app -- doesn't need to know secret key B - the write key).  Also, waits for card to be removed at the end of loop().