V2.2 of Station Config App
I noticed a cosmetic probem with the app and I cleaned it up.  Previously, there was a label for the new station type and another for the new listen type.  With the current GUI having one new type code and two buttons, these two text fields don't make sense.  There is now only one text field with the tyoe description that corresponds to the selected new code, regardless of whether the new code will be programmed to the device as a new station type or a new Listen code.