Update EZFLocalMemberCache_Proposal.md
I added in-line comments in square brackets:  [BG comment: ...].  I also added general comments and responses to JBS comments in the comments section, as BG:.

There is a lot of good stuff here, and a lot to talk through.  I'm not sure who else is interested in commenting and participating but we should have a comment deadline, followed shortly thereafter by a meeting (in person or virtual) where we can hash through all of the comments and create a clean, second version of this document.  From what I can see, we are all probably 90%+ in agreement here, so the next version of this document would be very clean; perhaps finalized.